Corrigir revisão do pr
- Arrumando ordem de criação de função
- Adicionando doc string
- Organizando arquivos
